+++ speedybot/__init__.py
@@ -1,6 +1,6 @@
 """SpeedyBot application package."""
 
-VERSION = "4.0.0"
+VERSION = "4.0.2"
 
 
 def install(core):